+++ docs/versionhistory.rst
@@ -3,6 +3,10 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
+**UNRELEASED**
+
+- Added support for ``typing.NoReturn``
+
 **2.8.0** (2020-06-02)
 
 - Added support for the ``Mock`` and ``MagicMock`` types (PR by prescod)